Score Manager update
Simple updatse where we are able to calculate the total score.

References Joker manager in order to   gain a list of the total additions from the jokers

Uses hand type to find the hand type of the played hand which will the return a base mult and chip value, to note I used AI to find a cleaner way to do if, else if, else if for the hand types.

Essentially just get base chips and mult, get jokerchips and mult then add or multiply them all together to get the final score and return, it will return to the game manager which will essentially be the main game loop manager utilizing all thhe created managers,